Server now hits Ouras endpoint and sends that data to a tag route which the front end now has access to which is tested and confirmed in the Home Component.
