+++ src/components/Home/Home.tsx
@@ -1,4 +1,5 @@
-import { useContext, useEffect } from "react";
+import { useContext, useEffect, useState } from "react";
+// import axios from "axios";
 import { GlobalContext } from "../../context/Provider";
 import Container from "../Container/Container";
 import HamburgerIcon from "../HamburgerIcon/HamburgerIcon";
@@ -7,10 +8,17 @@ import { HomeProps } from "../../types/dataTypes";
 
 const Home: React.FC<HomeProps> = ({ logout }) => {
   const { isMobileDisplay, setIsMobileDisplay } = useContext(GlobalContext);
+  const [tags, setTags] = useState<any>([]);
 
   useEffect(() => {
+    fetch(`http://localhost:8080/tags`)
+      .then((res) => res.json())
+      .then((data) => setTags(data.tagData.data));
+
     setIsMobileDisplay(false);
   }, [setIsMobileDisplay]);
+
+  console.log("tags", tags);
 
   const handleClickMobileDisplay = () => {
     setIsMobileDisplay(!isMobileDisplay);
